+++ ogma-core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for ogma-core
 
-## [1.X.Y] - 2026-09-12
+## [1.X.Y] - 2026-09-16
 
 * Replace adhoc function `mergeMaybe` with function from `base` (#516).
 * Fix malformed comments (#518).
@@ -20,6 +20,7 @@
 * Introduce default input format configuration files (#575).
 * Address HLint suggestions (#586).
 * Remove unused imports (#590).
+* Fix `mergeSpecs` dropping the first spec's external variables (#551).
 
 ## [1.15.0] - 2026-07-21